// Rho/src/types/schema.ts
// THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.

import { Document } from "@blockflow-labs/utils";

export class LiquidationBorrow {
  static entity = "LiquidationBorrow";
  static schema = {
    id: { type: "String", index: true },
    liquidator: "string",
    borrower: "string",
    repay_amount: "string",
    address: "string",
    seize_token: "string",
    rtoken_collateral: "string",
    block_timestamp: "string",
    transaction_hash: "string",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  };
}

export class Transfer {
  static entity = "Transfer";
  static schema = {
    id: { type: "String", index: true },
    from: "string",
    to: "string",
    amount: "string",
    block_timestamp: "string",
    transaction_hash: "string",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  };
}

import { String } from "@blockflow-labs/utils";

export interface ILiquidationBorrow extends Document {
  id: String;
  liquidator: string;
  borrower: string;
  repay_amount: string;
  address: string;
  seize_token: string;
  rtoken_collateral: string;
  block_timestamp: string;
  transaction_hash: string;
  blocknumber: String;
  entityId: String;
  instanceId: String;
  chainId: String;
}

export interface ITransfer extends Document {
  id: String;
  from: string;
  to: string;
  amount: string;
  block_timestamp: string;
  transaction_hash: string;
  blocknumber: String;
  entityId: String;
  instanceId: String;
  chainId: String;
}